# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for a Go Drunk Backend API, demonstrating DDD principles.",
        "title": "Go Drunk Backend API by DDD",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/2fa/disable": {
            "post": {
                "description": "Disable two-factor authentication for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Disable 2FA",
                "parameters": [
                    {
                        "description": "2FA disable request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFADisableReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/2fa/enable": {
            "post": {
                "description": "Enable two-factor authentication for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Enable 2FA",
                "parameters": [
                    {
                        "description": "2FA enable request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFAEnableReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/2fa/setup": {
            "post": {
                "description": "Setup two-factor authentication for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Setup 2FA",
                "parameters": [
                    {
                        "description": "2FA setup request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFASetupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns 2FA setup details",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_application_service_dto.TwoFASetupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/2fa/status": {
            "post": {
                "description": "Get two-factor authentication status for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get 2FA status",
                "parameters": [
                    {
                        "description": "2FA status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFAStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns 2FA status",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_application_service_dto.TwoFAStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/2fa/verify": {
            "post": {
                "description": "Verify a 2FA code for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify 2FA code",
                "parameters": [
                    {
                        "description": "2FA verification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFAVerifyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns verification result",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_application_service_dto.TwoFAVerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/change-password": {
            "post": {
                "description": "Change user password (requires current password)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Change password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.ChangePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    },
                    "401": {
                        "description": "Invalid current password",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/forgot-password": {
            "post": {
                "description": "Send password reset email to user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot password",
                "parameters": [
                    {
                        "description": "Forgot password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.ForgotPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate a user with username and password credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns authenticated user account information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format or validation errors",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    },
                    "401": {
                        "description": "Login failed due to invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Logout a user by invalidating their refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "parameters": [
                    {
                        "description": "Logout request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.LogoutReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_edynnt_veloras-api_internal_auth_application_service_dto.TwoFASetupResponse": {
            "type": "object",
            "properties": {
                "qr_code": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
...
        "github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFADisableReq": {
            "type": "object",
            "required": [
                "code",
                "user_id"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFAEnableReq": {
            "type": "object",
            "required": [
                "code",
                "secret",
                "user_id"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFASetupReq": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_edynnt_veloras-api_internal_auth_controller_dto.TwoFAStatusReq": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_edynnt_veloras-api_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "err": {},
                "errorFlag": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
...
}